+++ src/format.js
@@ -1,4 +1,5 @@
 (() => {
+const { escapeHtml } = window.EmbyMusicDomHelpers;
 function formatTicks(ticks) {
   if (!ticks) {
     return "--:--";
@@ -71,15 +72,6 @@ function coverClass(index) {
 
 function escapeHeaderValue(value) {
   return String(value).replace(/\\/g, "\\\\").replace(/"/g, '\\"');
-}
-
-function escapeHtml(value) {
-  return String(value)
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#039;");
 }
 
 window.EmbyMusicFormat = {